[ADD] Add Info Context Test
